requirements.txt
bcrypt==2.0.0
cffi==1.6.0
ecdsa==0.13
falcon==1.0.0
future==0.15.2
gevent==1.1.1
greenlet==0.4.9
gunicorn==19.6.0
peewee==2.8.1
pycparser==2.14
pycrypto==2.6.1
python-jose==0.6.2
python-mimeparse==1.5.2
six==1.10.0
